No CSS so flatten directory structure
