+++ app/odin-sequencer-react-ui/lib/components/MessageLog.jsx
