--- app/odin-sequencer-react-ui/lib/components/MessageLog/styles.css
@@ -1,33 +1,0 @@
-.ui-card {
-    margin-top: 20px;
-    border: 1px solid #ccc;
-    border-radius: 6px;
-    overflow: hidden;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05);
-    background-color: #fff;
-}
-
-.ui-card-header {
-    background-color: #f5f5f5;
-    padding: 8px 12px;
-    border-bottom: 1px solid #ddd;
-    font-weight: bold;
-    font-size: 1.1rem;
-}
-
-.ui-card-body {
-    padding: 8px;
-}
-
-.message-box {
-    height: 100%;
-    width: 100%;
-    overflow-y: auto;
-}
-
-.pre-scrollable {
-    background-color: #f7f7f7;
-    border: 1px solid black;
-    white-space: pre-wrap;
-    height: 600px
-}
