--- app/odin-sequencer-react-ui/lib/components/SequenceButtons/styles.css
@@ -1,7 +1,0 @@
-.button-row {
-    display: flex;
-    justify-content: flex-end;
-    width: 90%;
-    gap: 50px;
-    margin: 0 auto;
-}
