+++ app/odin-sequencer-react-ui/lib/components/SequenceModule.jsx
@@ -1,0 +1,22 @@
+import Accordion from 'react-bootstrap/Accordion'
+import CardRow from './CardRow'
+import { Row } from 'react-bootstrap'
+
+/* Constructs Accordion layer for each sequence module, then intialises the CardRow to display sequences. */
+
+const SequenceModule = ({sequences, header, executionPanelRef, setAbortDisabled, sequencer_endpoint}) => {
+  return (
+    <Accordion.Item eventKey={header}>
+      <Accordion.Header>{header}</Accordion.Header>
+      <Accordion.Body>
+        <Row>
+          <CardRow sequences={sequences} row_title={header} executionPanelRef={executionPanelRef} setAbortDisabled={setAbortDisabled} sequencer_endpoint={sequencer_endpoint}></CardRow>
+        </Row>
+      </Accordion.Body>
+    </Accordion.Item>
+  )
+}
+
+//Look into odin-react title card
+
+export default SequenceModule
