--- app/odin-sequencer-react-ui/lib/components/SequenceTable/styles.css
@@ -1,128 +1,0 @@
-.alert-box {
-    position: sticky;
-    z-index: 1040;
-    padding-top: 18px;
-    padding-left: 12px;
-    padding-right: 12px;
-}
-
-.flex-container {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-.left {
-    width: 58%;
-    min-width: 400px;
-    padding-right: 12px;
-    padding-left: 24px;
-}
-
-.ui-card {
-    margin-top: 20px;
-    border: 1px solid #ccc;
-    border-radius: 6px;
-    overflow: hidden;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05);
-    background-color: #fff;
-}
-
-.ui-card-header {
-    background-color: #f5f5f5;
-    padding: 8px 12px;
-    border-bottom: 1px solid #ddd;
-    font-weight: bold;
-    font-size: 1.1rem;
-}
-
-.ui-card-body {
-    padding: 8px;
-}
-
-.right {
-    width: 42%;
-    padding-left: 24px;
-    padding-right: 12px;
-}
-
-.card-title {
-    font-size: 0.9rem !important;
-    text-align: center;
-}
-
-.card-body {
-    border-radius: 0px;
-    padding: 4px !important;
-    width: 184px;
-    height: 55px;
-}
-
-.card {
-    max-width: 184px;
-}
-
-.col {
-    max-width: 184px !important;
-    padding-left: 0px !important;
-    padding-right: 0px !important;
-}
-
-.btn-width {
-    width: 58px;
-}
-
-.card-new{
-    max-width: 170px;
-    min-width: 210px;
-}
-
-.col-card {
-    padding-top: 4px;
-    max-width: 220px;
-}
-
-.sticky{
-    height: 100%;
-    width: 48%;
-    float:right;
-    position: fixed;
-}
-
-.custom-row-border{
-    margin-right:0px;
-    margin-left: 0px;
-}
-
-.row{
-    margin-left: 0 !important;
-    margin-right: 0 !important;
-}
-
-.accordion-button{
-    padding-left: 0.75rem !important;
-    padding-right: 0.75rem !important;
-    padding-top: 0.5rem !important;
-    padding-bottom: 0.5rem !important;
-}
-
-.accordion-body {
-    padding: 0px !important;
-}
-
-.card-header {
-    color: #007bff;
-}
-
-.col-sm-7 {
-    width: 100% !important;
-}
-
-@media (max-width: 768px) {
-    .left, .right {
-        width: 100%;
-    }
-}
-
-.alert-bar {
-    padding-top: 18px;
-}
